pyproject.toml
[tool.poetry]
name = "rag-service"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) service with FastAPI, OpenAI, Qdrant"
authors = []
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = { extras = ["standard"], version = "^0.30" }
openai = "^1.40"
qdrant-client = { extras = ["async"], version = "^1.10" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0" }
asyncpg = "^0.29"
alembic = "^1.13"
redis = { extras = ["hiredis"], version = "^5.0" }
pydantic-settings = "^2.4"
python-multipart = "^0.0.12"
pypdf = "^4.3"
python-docx = "^1.1"
structlog = "^24.4"
python-jose = { extras = ["cryptography"], version = "^3.3" }
aiofiles = "^24.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
httpx = "^0.27"
pytest-cov = "^5.0"
ruff = "^0.6"
mypy = "^1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
